+++ src/index.full.ts
@@ -1,0 +1,3 @@
+export * from './index.lite';
+
+export { plotlib, graphlib, geolib, threedlib } from './lib';
